Add files via upload
Title: Simple Calculator

Description:
The Simple Calculator project is a basic web-based calculator that allows users to perform arithmetic operations such as addition, subtraction, multiplication, and division. It provides a simple and user-friendly interface for performing calculations. Users can input numbers and perform operations by clicking on the calculator buttons.

Features:

Addition, subtraction, multiplication, and division operations.
Responsive design for a better user experience on different devices.
Clear button to reset the input and start a new calculation.
Technologies Used:

HTML for structuring the web page.
CSS for styling and layout.
JavaScript for handling user interactions and performing calculations.
How to Use:

Enter numerical values using the calculator buttons.
Click on the operation buttons (+, -, *, /) to perform calculations.
Press the "=" button to see the result of the calculation.
Use the "C" button to clear the input and start a new calculation.